move application destination info to bootstrap-application-base
