+++ generators/base/generator-base-todo.mjs
@@ -31,7 +31,6 @@ import { fileURLToPath } from 'url';
 import { formatDateForChangelog, normalizePathEnd, createJHipster7Context } from './support/index.mjs';
 import { packageJson } from '../../lib/index.mjs';
 import PrivateBase from './generator-base-definitions.mjs';
-import NeedleApi from '../needle-api.mjs';
 import commonOptions from './options.mjs';
 import { detectLanguage, loadLanguagesConfig } from '../languages/support/index.mjs';
 import { getDBTypeFromDBValue, calculateDbNameWithLimit, hibernateSnakeCase } from '../server/support/index.mjs';
@@ -855,42 +854,6 @@ export default class JHipsterBaseGenerator extends PrivateBase {
         }
       );
     }
-  }
-
-  /**
-   * Prints a JHipster logo.
-   */
-  printJHipsterLogo(cwd = this.destinationPath()) {
-    this.logger.log(chalk.white(`Application files will be generated in folder: ${chalk.yellow(cwd)}`));
-    if (process.cwd() === this.getUserHome()) {
-      this.logger.log(chalk.red.bold('\n️⚠️  WARNING ⚠️  You are in your HOME folder!'));
-      this.logger.log(
-        chalk.red('This can cause problems, you should always create a new directory and run the jhipster command from here.')
-      );
-      this.logger.log(chalk.white(`See the troubleshooting section at ${chalk.yellow('https://www.jhipster.tech/installation/')}`));
-    }
-    this.logger.log(
-      chalk.green(' _______________________________________________________________________________________________________________\n')
-    );
-    this.logger.log(
-      chalk.white(`  Documentation for creating an application is at ${chalk.yellow('https://www.jhipster.tech/creating-an-app/')}`)
-    );
-    this.logger.log(
-      chalk.white(
-        `  If you find JHipster useful, consider sponsoring the project at ${chalk.yellow('https://opencollective.com/generator-jhipster')}`
-      )
-    );
-    this.logger.log(
-      chalk.green(' _______________________________________________________________________________________________________________\n')
-    );
-  }
-
-  /**
-   * @private
-   * Return the user home
-   */
-  getUserHome() {
-    return process.env[isWin32 ? 'USERPROFILE' : 'HOME'];
   }
 
   /**
