+++ generators/server/generator.mjs
@@ -201,12 +201,6 @@ export default class JHipsterServerGenerator extends BaseApplicationGenerator {
 
   get initializing() {
     return this.asInitializingTaskGroup({
-      displayLogo() {
-        if (this.logo) {
-          this.printJHipsterLogo();
-        }
-      },
-
       validateJava() {
         if (!this.options.skipChecks) {
           this.checkJava();
